fixed：关闭爆品弹窗刷新列表

+++ src/views/mall/explodeGoods/explodeGoods.js
@@ -188,6 +188,7 @@ export default {
             type: 'success'
           })
           this.getAllGoodsLists()
+          this.fetchData()
         }
       })
     },
@@ -196,11 +197,29 @@ export default {
       console.log(row)
       let delGoodsId = []
       delGoodsId.push(row.id)
-      deleteExplodeGoods(delGoodsId).then(res => {
-        if (res.state === 1) {
-          this.fetchData()
+      this.$confirm('', '确定移除？', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消'
+      }).then(() => {
+        let params = {
+          'id': row.id
         }
+        deleteExplodeGoods(delGoodsId).then(res => {
+          if (res.state === 1) {
+            this.$message({
+              message: '移除成功',
+              type: 'success'
+            })
+            this.fetchData()
+          }
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '移除删除'
+        })
       })
+
     },
     // 批量移除
     handleSelectionChange(value) {
